Structure mdlint migrate for multiple sources via --from
Adds an explicit --from <SOURCE> flag (defaulting to and currently only
accepting markdownlint-cli2) and moves the markdownlint-cli2-specific
parsing into its own src/migrate/markdownlint_cli2/ submodule, so the
CLI shape and docs read as multi-source even though only one source is
implemented today. A future source only needs a new module and
MigrateFrom variant, mirroring how `uv export` exposed --format long
before it supported more than one. Also corrects the README's outdated
"does not execute JavaScript" claim now that Node-eval is in place.
