+++ CLAUDE.md
@@ -25,7 +25,8 @@ src/
   fix/                   # Auto-fix framework
   formatter/             # Canonical markdown rewriter (mdlint format)
   format/                # Output formatters (default, JSON, JUnit, SARIF)
-  migrate/               # markdownlint-cli2 config -> mdlint.toml (mdlint migrate)
+  migrate/               # config migration to mdlint.toml (mdlint migrate), one submodule per
+                          # source tool selected via --from (markdownlint_cli2/ today)
   logger/                # Log level handling
   error.rs / types.rs    # Shared types and error definitions
 ```
