+++ python/README.md
@@ -138,25 +138,31 @@ Options:
 
 ### mdlint migrate
 
-Migrate a `markdownlint-cli2` configuration to `mdlint.toml`.
+Migrate a configuration from another Markdown tool to `mdlint.toml`. Currently supported sources (`--from`):
+`markdownlint-cli2` (default).
 
 ```text
 Usage: mdlint migrate [OPTIONS] [INPUT]
 
 Arguments:
-  [INPUT]                          Path to the markdownlint-cli2 config to migrate (auto-detected if omitted)
+  [INPUT]                          Path to the configuration to migrate (auto-detected if omitted)
 
 Options:
-      --output <PATH>              Output path for the generated config [default: mdlint.toml]
-      --force                      Overwrite the output file if it already exists
-      --dry-run                    Print the generated config without writing it
-  -h, --help                       Print help
+      --from <SOURCE>               Tool to migrate the configuration from [default: markdownlint-cli2]
+      --output <PATH>               Output path for the generated config [default: mdlint.toml]
+      --force                       Overwrite the output file if it already exists
+      --dry-run                     Print the generated config without writing it
+  -h, --help                        Print help
 ```
+
+#### markdownlint-cli2
 
 Supports `.markdownlint-cli2.{json,jsonc,yaml,yml}` and standalone `.markdownlint.{json,jsonc,yaml,yml}` rule
 configs. Rule names and their common aliases (e.g. `line-length` for `MD013`) are both recognized. `.cjs`/`.mjs`
-configs are handled on a best-effort basis (mdlint does not execute JavaScript) — if a config uses variables,
-function calls, or computed values, migration fails with a message asking you to export the config with
+configs are evaluated with a Node.js runtime when one is found on `PATH` (the same thing `markdownlint-cli2` itself
+would do when loading them), correctly resolving `require()`, spread syntax, and computed values. If Node isn't
+available, mdlint falls back to a best-effort text scrape and warns that dynamic values may not have been resolved;
+if a config can't be parsed either way, migration fails with a message asking you to export it with
 `console.log(JSON.stringify(config))` and migrate the resulting JSON file instead. Rules with no mdlint
 implementation, and cli2-specific fields with no mdlint equivalent (`globs`, `customRules`, `outputFormatters`),
 are skipped with a warning rather than failing the migration.
@@ -194,11 +200,11 @@ mdlint check --config path/to/mdlint.toml
 # ignore all config files
 mdlint check --no-config
 
-# migrate a markdownlint-cli2 config to mdlint.toml
+# migrate a markdownlint-cli2 config to mdlint.toml (auto-detected)
 mdlint migrate
 
 # migrate a specific config file to a custom output path
-mdlint migrate .markdownlint-cli2.jsonc --output mdlint.toml
+mdlint migrate --from markdownlint-cli2 .markdownlint-cli2.jsonc --output mdlint.toml
 ```
 
 ## Configuration
